requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
black==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.1.4
django-browser-reload==1.6.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
filelock==3.6.0
flake8 @ file:///opt/conda/conda-bld/flake8_1648129545443/work
fonttools==4.34.4
gunicorn==20.1.0
idna==3.4
importlib-metadata @ file:///C:/ci/importlib-metadata_1648562621412/work
ipykernel==6.9.2
ipython==8.1.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.1.2
jupyter-core==4.9.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe @ file:///opt/conda/conda-bld/mccabe_1644221741721/work
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.0.10
pandas==1.5.2
parso==0.8.3
pathspec==0.10.2
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.1
prompt-toolkit==3.0.28
psutil==5.9.0
pure-eval==0.2.2
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1636635402688/work
pycparser==2.21
pyflakes @ file:///tmp/build/80754af9/pyflakes_1636644436481/work
Pygments==2.11.2
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
pywin32==303
pyzmq==22.3.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stack-data==0.2.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.13.4
wcwidth==0.2.5
wincertstore==0.2
zipp @ file:///C:/ci/zipp_1652273994994/work
